New client setup with Parcel
