+++ client/.babelrc
@@ -1,14 +1,3 @@
 {
-  "presets": [
-    ["@babel/env", {
-      "targets": {
-        "browsers": [
-          ">0.25%",
-          "not ie 11",
-          "not op_mini all"
-        ]
-      },
-      "modules": false
-    }]
-  ],
+  "presets": ["env", "react"]
 }
