+++ client/index.html
@@ -1,8 +1,10 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <meta charset="utf-8">
+    <title>GameOver</title>
   </head>
   <body>
+    <div id="app"></div>
+    <script src="./src/index.js"></script>
   </body>
 </html>
