+++ client/src/game/assets/backgrounds/background.png
