+++ client/src/game/assets/backgrounds/platform.png
