+++ client/src/game/assets/backgrounds/sky.png
