+++ client/src/game/assets/maps/inca2.tmx
