+++ client/src/game/assets/maps/inca_back2.png
