+++ client/src/game/assets/maps/inca_front.png
