+++ client/src/game/assets/sprites/bomb.png
