+++ client/src/game/assets/sprites/dude.png
