+++ client/src/game/assets/sprites/star.png
